--- components/identity-core/org.wso2.carbon.identity.core/src/main/java/org/wso2/carbon/identity/core/util/MDCUtils.java
@@ -1,51 +1,0 @@
-/*
- *  Copyright (c) 2024, WSO2 LLC. (http://www.wso2.com).
- *
- *  WSO2 LLC. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- *
- */
-
-package org.wso2.carbon.identity.core.util;
-
-import org.slf4j.MDC;
-
-import java.util.Map;
-
-/**
- * Util class for MDC.
- */
-public class MDCUtils {
-
-    public static Runnable wrapWithMdcContext(Runnable task) {
-        // Save the current MDC context
-        Map<String, String> contextMap = MDC.getCopyOfContextMap();
-        return () -> {
-            setMDCContext(contextMap);
-            try {
-                task.run();
-            } finally {
-                // Once the task is complete, clear MDC
-                MDC.clear();
-            }
-        };
-    }
-
-    public static void setMDCContext(Map<String, String> contextMap) {
-        MDC.clear();
-        if (contextMap != null) {
-            MDC.setContextMap(contextMap);
-        }
-    }
-}
